Move the Notifier to a dedicated package
The Notifier is a high-level abstraction that does not leak any apparmor
implementation details directly. As such it should be in a different
package.

Some methods were adjusted to better accommodate the new package name.
