+++ prompting/notifier/notifier_test.go
@@ -1,8 +1,12 @@
-package apparmor_test
+package notifier_test
 
 import (
+	"testing"
+
 	. "gopkg.in/check.v1"
 )
+
+func Test(t *testing.T) { TestingT(t) }
 
 type notifierSuite struct{}
